--- a/package.json
+++ b/package.json
@@ -210,7 +210,6 @@
     "prop-types": "15.8.1",
     "qhistory": "1.1.0",
     "qs": "6.10.3",
-    "raf": "3.4.1",
     "rc-tooltip": "5.1.1",
     "react": "16.14.0",
     "react-autosuggest": "10.1.0",