--- a/package.json
+++ b/package.json
@@ -230,7 +230,6 @@
     "redux": "4.1.2",
     "redux-logger": "3.0.6",
     "redux-saga": "0.16.2",
-    "regenerator-runtime": "0.13.9",
     "response-time": "2.3.2",
     "schema-utils": "4.0.0",
     "serialize-javascript": "6.0.0",