--- a/src/amo/polyfill.js
+++ b/src/amo/polyfill.js
@@ -1,3 +1,2 @@
-import 'raf/polyfill';
 import 'isomorphic-fetch';
 import 'focus-visible';