--- a/src/amo/polyfill.js
+++ b/src/amo/polyfill.js
@@ -1,5 +1,3 @@
-import 'core-js/stable';
-import 'regenerator-runtime/runtime';
 import 'raf/polyfill';
 import 'isomorphic-fetch';
 import 'focus-visible';