--- a/webpack.l10n.config.babel.js
+++ b/webpack.l10n.config.babel.js
@@ -1,6 +1,4 @@
 /* eslint-disable no-console, import/no-extraneous-dependencies */
-import fs from 'fs';
-
 import chalk from 'chalk';
 import webpack from 'webpack';
 
@@ -13,9 +11,8 @@ if (process.env.NODE_ENV !== 'production') {
   process.exit(1);
 }
 
-const babelrc = fs.readFileSync('./.babelrc');
-const babelrcObject = JSON.parse(babelrc);
-const babelPlugins = babelrcObject.plugins || [];
+const babelConfig = require('./babel.config.js');
+const babelPlugins = babelConfig.plugins || [];
 
 // Create UTC creation date in the correct format.
 const potCreationDate = new Date()
@@ -57,7 +54,7 @@ const babelL10nPlugins = [
 ];
 
 const babelOptions = {
-  ...babelrcObject,
+  ...babelConfig,
   plugins: babelPlugins.concat(babelL10nPlugins),
 };