Related to #6192

This PR includes:
- update of `browserslist`'s `caniuse-lite` DB to reflect latest browser targets as defined in `.browserslistrc`.
- fine-tuning of the Babel configuration, according to the current state of Babel related dependencies & browser targets defined in `.browserslistrc`, and appropriate removal of no-longer relevant `core-js` / `regenerator-runtime` polyfill explicit imports.
- removal of probably unused / unnecessary `raf` polyfill.
- migration from `.babelrc` file to `babel.config.js` which enables commenting the configuration file.
- dependencies cleanup.
- upgrade to latest Babel dependencies.

# Details

- ref 24c767259d14d292b4603fd750a575828ad17ac4
- command `yarn build`

## Target browser changes

```diff
Target browser changes:
- and_chr 96
+ and_chr 102
- and_ff 95
+ and_ff 101
- android 96
+ android 101
- chrome 97
- chrome 96
- chrome 95
- chrome 94
+ chrome 102
+ chrome 101
+ chrome 100
- edge 97
- edge 96
+ edge 102
+ edge 101
- firefox 95
- firefox 94
+ firefox 101
+ firefox 100
- ios_saf 15.2
+ ios_saf 15.5
+ ios_saf 15.4
+ ios_saf 15.2-15.3
- opera 82
- opera 81
+ opera 86
+ opera 85
- safari 15.2
- safari 15.1
- safari 14.1
- safari 13.1
+ safari 15.5
+ safari 15.4
- samsung 14.0
+ samsung 16.0
```
## JavaScript output files

65 .js files

### Before

```
size: 5461336 bytes
size (pretty): 5.46 MB
parse ~ (espree): 1065.69 ms
```

### After

```
size: 5399015 bytes
size (pretty): 5.4 MB
parse ~ (espree): 982.72 ms
```

### Diff

```
size: 62321 bytes 
size (pretty): 62.3 kB bytes
relative: 1.14%
```

## Main bundle

- `amo-[contenthash].js`

### Before

```
file: `amo-e032875f63792ae6d8d0.js`
size: 1266124 bytes
size (pretty): 1.27 MB
size brotli 9: 332479 bytes
size brotli 9 (pretty): 332 kB
parse ~ (espree): 806.95 ms
```

### After

```
file: `amo-1f0b50817dc57a6675e2.js`
size: 1204086 bytes
size (pretty): 1.2 MB
size brotli 9: 318328 bytes
size brotli 9 (pretty): 318 kB
parse ~ (espree): 725.91 ms
```

### Size gains (uncompressed)

```
size gain: 62038 bytes 
size gain (pretty): 62 kB
size gain (relative): 4.89%
```

### Size gains (brotli 9)

```
size gain: 14151 bytes
size gain (pretty): 14.2 kB
size gain (relative): 4.25%
```

### Parsing time ~ (espree)

```
parsing time ~ gain:  81.04 ms
parsing time ~ gain (relative): 10.04%
```

## Bundle diff (prettified)

### Template literals

```diff
<               n.info("[install] Adding listener for ".concat(t)),
---
>               n.info(`[install] Adding listener for ${t}`),
```

### Tagged templates

```diff
<                   (0, o.f9)(
<                     r ||
<                       ((n = [
<                         'Key "',
<                         '" was not found in clientConfig. Check the\n          key has been added to clientConfigKeys',
<                       ]),
<                       i || (i = n.slice(0)),
<                       (r = Object.freeze(
<                         Object.defineProperties(n, {
<                           raw: { value: Object.freeze(i) },
<                         })
<                       ))),
---
>                   (0, r.f9)(
>                     o ||
>                       (o = i`Key "${0}" was not found in clientConfig. Check the
>           key has been added to clientConfigKeys`),
```

### Spreading props + static class properties

```diff
<       function ki() {
<         return (ki =
<           Object.assign ||
<           function (e) {
<             for (var t = 1; t < arguments.length; t++) {
<               var n = arguments[t];
<               for (var r in n)
<                 Object.prototype.hasOwnProperty.call(n, r) && (e[r] = n[r]);
<             }
<             return e;
<           }).apply(this, arguments);
<       }
<       function Ai(e, t, n) {
<         return (
<           t in e
<             ? Object.defineProperty(e, t, {
<                 value: n,
<                 enumerable: !0,
<                 configurable: !0,
<                 writable: !0,
<               })
<             : (e[t] = n),
<           e
<         );
<       }
<       var Ni = () => (e) => {
---
>       var Ai = () => (e) => {
48848c45790
<             super(e, t), Ai(this, "i18n", void 0), (this.i18n = t.i18n);
---
>             super(e, t), (this.i18n = t.i18n);
48851c45793,45796
<             return r.createElement(e, ki({ i18n: this.i18n }, this.props));
---
>             return r.createElement(
>               e,
>               Object.assign({ i18n: this.i18n }, this.props)
>             );
48855,48856c45800,45801
<           Ai(t, "contextTypes", { i18n: w().object }),
<           Ai(t, "displayName", "Translate(".concat(Nt(e), ")")),
---
>           (t.contextTypes = { i18n: w().object }),
>           (t.displayName = `Translate(${Nt(e)})`),
```

### Class properties

```diff
<         _y = (0, r.forwardRef)(Ey);
<       class xy extends r.Component {
<         constructor(...e) {
<           super(...e),
<             (function (e, t, n) {
<               t in e
<                 ? Object.defineProperty(e, t, {
<                     value: n,
<                     enumerable: !0,
<                     configurable: !0,
<                     writable: !0,
<                   })
<                 : (e[t] = n);
<             })(this, "container", void 0);
<         }
---
>         Lg = (0, r.forwardRef)(Dg);
>       class Mg extends r.Component {
63601c59595
<             s = "".concat(t || "", "TooltipMenu");
---
>             s = `${t || ""}TooltipMenu`;
```

## Build performance

```
before: 80.83s
after: 74.41s
gain: 6.42s
gain relative: 7.9%
```